refactor: 안 쓰이는 부분 지우기
