+++ src/client/Client.cpp
@@ -4,7 +4,7 @@
 
 #include <iostream>
 
-#include "command/returncode.hpp"
+#include "returncode.hpp"
 #include "server/Server.hpp"
 #include "util/algorithm/algorithm.hpp"
 #include "util/color.hpp"
